整合PR #1823
