+++ src/MiniProgram/Application.php
@@ -41,7 +41,6 @@ use EasyWeChat\Kernel\ServiceContainer;
  * @property \EasyWeChat\MiniProgram\SubscribeMessage\Client    $subscribe_message
  * @property \EasyWeChat\MiniProgram\RealtimeLog\Client         $realtime_log
  * @property \EasyWeChat\MiniProgram\Search\Client              $search
- * @property \EasyWeChat\MiniProgram\Live\Client                $live
  * @property \EasyWeChat\MiniProgram\Broadcast\Client           $broadcast
  */
 class Application extends ServiceContainer
@@ -69,7 +68,6 @@ class Application extends ServiceContainer
         SubscribeMessage\ServiceProvider::class,
         RealtimeLog\ServiceProvider::class,
         Search\ServiceProvider::class,
-        Live\ServiceProvider::class,
         Broadcast\ServiceProvider::class,
         // Base services
         BasicService\Media\ServiceProvider::class,
