--- src/MiniProgram/Live/ServiceProvider.php
@@ -1,33 +1,0 @@
-<?php
-
-/*
- * This file is part of the overtrue/wechat.
- *
- * (c) overtrue <i@overtrue.me>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace EasyWeChat\MiniProgram\Live;
-
-use Pimple\Container;
-use Pimple\ServiceProviderInterface;
-
-/**
- * Class ServiceProvider.
- *
- * @author onekb <1@1kb.ren>
- */
-class ServiceProvider implements ServiceProviderInterface
-{
-    /**
-     * {@inheritdoc}.
-     */
-    public function register(Container $app)
-    {
-        $app['live'] = function ($app) {
-            return new Client($app);
-        };
-    }
-}
